Safari-611.1.10/LayoutTests/fast/dom/Document/replaceChild-null-oldChild.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description('Test behavior of Document.replaceChild() when oldChild is null.');

shouldThrow('document.replaceChild(document.firstChild, null)');
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
